LazyRefreshable implementation to avoid `map` overhead
